last clicked book to store & Selected Books Sidebar to component for testing
